requirements.txt
requests
playwright
python-dotenv
Pillow
numpy
opencv-python-headless
